Detect parenthesized SPDX identifier or exact name match in liberal mode
E.g. from "Some freeform text here (Mozilla Public License 2.0)" we'll
pick up that "Mozilla Public License 2.0" is the exact name for MPL-2.0
in the SPDX data.

E.g. from "Blah Blah Blah (MIT)" we'll recognize that "MIT" is a valid
SPDX identifier in the SPDX data.

However, if it's a longer text that contains multiple parenthesized
license names or identifiers, even exact matches, we won't make any
assumption about whether it's an "AND" or an "OR".